Add targets for terraform lockfiles (#21099)
Terraform lockfiles were magically pulled in with a pathglob. This meant
that `--changed-since` had no visibility on changes in the lockfile.
This MR adds a synthetic target for terraform lockfiles, wires it into
dep inference, and uses its Sources to pull it in.

The key reason to use a synthetic target instead of a normal target with
tailor is that the lockfile _must_ be called `.terraform.lock.hcl` and
must be in the same directory as the root module of a Terraform
deployment. So any such file _must_ be a Terraform lockfile, and no
other file _can_ be.

closes #14790