Helm: warn on missing docker image (#21076)
Add a toggle for warning if something that will be used as an image
looks like a target but does not exist. Also adds an option for known
3rd-party images (and the option to disable it).

This allows us to ensure that an image is valid. Obviously, if it's a
valid `docker_image` target, it's valid. Otherwise, the value being
passed to an image could fit into several categories, which we can warn
about:

- a 3rd party image : we check that this is a known image, so
  `busybox:latest` isn't interpreted as a target
- an image that was produced by a `docker_image` target, but taken from
  the registry : `myproject/docker:1.0.0`. we treat is as a 3rd party
  chart, which isn't accurate but has the same outcome.
- something that is obviously a Pants target but that isn't found :
  `//myproject:oops`, which we can warn must be a typo
- targets of the wrong type : `myproject:files`

closes to #18346