tests: add experimental-report-test-result-info option to the test goal (#21171)
A new option `--report-test-result-source` is added to the `[test]`
config section. Enabling this option will
produce a file on disk with information that would tell you where the
results of tests come from. The tests might
have been executed locally or remotely, but they might have been
retrieved from the local or remote cache,
or be memoized. 

Knowing where the test results come from may be useful
when evaluating the efficiency of the
caching strategy and the nature of the changes in the source code that
may lead to frequent cache invalidations.