deprecate `--export-py-hermetic-scripts` in favor of new `--export-py-non-hermetic-scripts-in-resolve` option (#21181)
Deprecate the `--export-py-hermetic-scripts` option in favor of the new
`--export-py-non-hermetic-scripts-in-resolve` option so that the
hermetic script logic can be configured on a per-resolve basis. The
behavior for the new option is inverted in a sense from the old option
since if we named the option `--export-py-hermetic-scripts-in-resolve`
then the default option value of an empty list would mean that resolve
did not use hermetic scripts by default. Thus, the new option acts
instead as an opt-in to use non-hermetic scripts.