Allow customizing semgrep configurations; correct rule matching glob (#21126)
Three things:

Allow configuring the config location; because of how the digest is 
built, running `semgrep` via pants doesn't actually  allow you to 
configure this by passing `--config`, like the tool does. There is now
a `config_name` option which will be auto-discovered and hierarchically
applied to each relevant partition. This will potentially evolve into
a `config_paths` option which will allow users to specify a list of
globs.

`semgrep` searches for rules recursively within a rules directory, so 
reflect that in the globs. This isn't obvious in the documentation, but
it's the way it behaves, and also is explicit in the code.

https://github.com/semgrep/semgrep/blob/45699ace83dd61f19d6ea1e7452f32a0c1996735/cli/src/semgrep/config_resolver.py#L402


Finally, I flattened the glob for `.semgrepignore`. Currently, the glob
recursively matches all matching file names, but ultimately `semgrep` 
just reads only what's in the current dir. In my testing of
pants+semgrep
I did indeed notice the same behavior, that the `.semgrepignore`s in
lower
dirs were not being respected. An integration test confirmed the
same.[1]
Supporting multi-level `.semgrepignores` is going to be
a significant effort, since it would involve actually parsing and
interpreting
the config files, and also navigating the pitfalls of managing overlays
of
ignore patterns. Instead, I chose to force it to being a singular file 
path.

Here're some historical discussions about .semgrepignore:
https://github.com/huonw/pants/pull/1,
https://github.com/pantsbuild/pants/pull/18593#discussion_r1184441437

https://github.com/semgrep/semgrep/blob/45699ace83dd61f19d6ea1e7452f32a0c1996735/src/targeting/Semgrepignore.ml#L124
https://semgrep.dev/docs/ignoring-files-folders-code#reference-summary

[1] The integration test which proves hierarchical semgrepignore doesn't
work:
```py
def test_semgrepignore_not_partitioned(rule_runner: RuleRunner) -> None:
    rule_runner.write_files({
        f"{DIR}/BUILD": "",
        f"{DIR}/.semgrep.yml": RULES,
        f"{DIR}/.semgrepignore": FILE,
        
        f"{DIR}/a/.semgrepignore": FILE,
        f"{DIR}/a/BUILD": SINGLE_FILE_BUILD,
        f"{DIR}/a/{FILE}": BAD_FILE,
    })
    tgt = rule_runner.get_target(Address(f"{DIR}/a", target_name="f"))

    # run with these two semgrepignores in place
    results = run_semgrep(rule_runner, [tgt])
    assert len(results) == 1
    result = results[0]
    # doesn't get ignored
    assert "Ran 1 rule on 1 file: 1 finding" in result.stderr

    # run with a root semgrepignore
    rule_runner.write_files({
        ".semgrepignore": FILE,
    })
    results = run_semgrep(rule_runner, [tgt])
    assert len(results) == 1
    result = results[0]
    # finally, it's ignored
    assert "Ran 1 rule on 0 files: 0 findings" in result.stderr
```

---------

Co-authored-by: Huon Wilson <wilson.huon@gmail.com>