add `path_env_modify` field to `adhoc_tool` and `shell_command` to control how PATH is augmented (#21138)
As described in https://github.com/pantsbuild/pants/issues/21136, Pants
currently has a bug where any `PATH` value in the `extra_env_vars` field
of the `shell_command` target type is ignored and the `PATH` is set to
only the binary shim directory.

This PR builds on top of https://github.com/pantsbuild/pants/pull/21185
to allow the user to configure how the `PATH` is augmented. It adds a
new `path_env_modify` field on `adhoc_tool` and `shell_command` targets
which configures whether the shims are prepended (`prepend`), appended
(`append`), or ignored (`off`) to the existing `PATH` value.