Support recursive rule call-by-name. (#21174)
Previously, a call-by-name was detected by looking for 
the `rule` attr on the callable. However this attr can only
be added after the rule and all its calls have finished
processing, which precludes recursive call-by-name.

Now we set a `rule_id` attr on the callable, which we
can do before the calls have been processed, and use
that for detection and naming.