use entire URL string as cache key for download intrinsic rule (Cherry pick of #21215) (#21244)
As reported in https://github.com/pantsbuild/pants/issues/21163, query
parameters apparently do not form part of the cache key when downloading
files. Thus, while the query will be used on the first download from a
URL, subsequent changes to just the query portion of the URL do not
invalidate the cache and do not re-download the file. Thus, the
intrinsic would return stale cached content.

Switch the cache key for downloaded files to use the entire URL and not
just the path. This will pick up query string changes and host name
changes. Fixes the issue in manual testing.

Also adds a regression test for the bug.