Enable MockGet when using call-by-name (#21144)
The scaffolding required to use `MockGet` as advertised required `Call`
exposing some variables that were kept internal compared to `Get`.

I think this a good compromise to keep the migration to call-by-name
smooth. Especially since plugins out in the wild might utilize `MockGet`
more than pants does internally as it is advertized in the
[docs](https://www.pantsbuild.org/2.22/docs/tutorials/testing-plugins#unit-testing-for-rules),
and plugin authors probably want their plugins to be fast, too.