lookup the correct EnvironmentTarget for adhoc_tool (#21253)
One of the `adhoc_tool`-related rules was not looking up the correct
`EnvironmentTarget` since it "changes into" a different environment
inside the rule. Thus, the existing code taking the `EnvironmentTarget`
as a parameter meant that the engine was supplying the
`EnvironmentTarget` for the "outer" environment and not the one switched
to for the `adhoc_tool` execution.

I discovered this bug while trying to fix
https://github.com/pantsbuild/pants/issues/21137 and wondering why my
modifications to `EnvironmentTarget` for workspace environments were
being ignored.