Terraform multi-platform lockfiles (#21247)
Terraform lockfiles include the hashes for only a subset of platforms.
On an upgrade, it will only use the current platform. The intention is
that every other platform will run `terraform init`, and this will
modify the lockfile. We generate lockfiles statically and don't allow
them to be modified, so this expectation doesn't hold. Terraform
supports generating a lockfile with multiple providers with the
`terraform providers lock` command.

This MR runs this command with a list of providers forwarded to it.

Closes #21229