jvm resources action is not reproducible (#21259)
While we are trying to zip the file without extra data and same creation
date, we noticed in CI that the action can produce different digests.
This commit adds another strip jar action after zipping the resources.

It would be great if this can be cherry-picked to 2.20.X and above