adhoc/shell: capture files relative to root of sandbox for in-workspace executions (#21254)
As described in https://github.com/pantsbuild/pants/issues/21137, there
is an inconsistent developer experience currently for users of the
in-workspace execution support with the `adhoc_tool` / `shell_command`
target types. Specifically, the `output_files` and `output_directories`
fields have the working directory prepended to them. This is intentional
for most adhoc_tool / shell_command executions since they occur in a
sandboxed environment and the execution's working directory and the
capture base directory are the same.

For in-workspace executions, however, those directories are different
and the `{chroot}` replacement refers to a separate temporary directory
and not any directory inside the workspace. Users of the in-workspace
execution support have had to prepend `..` path elements to
`output_files` / `output_directories` values in order to remove the
prepended working directory path.

This PR fixes the issue by having Pants interpret `output_files` /
`output_directories` relative to the sandbox root with no prepending of
the working directory for in-workspace executions.