Allow int values for float options in native config file parser (Cherry-pick of #21265) (#21273)
This adjusts the native parser handling of a TOML config file to allow
an integer value to satisfy a float option, e.g. `123` and `123.0` are
now equivalent, previously `123` was rejected. This makes config file
parsing match env var and arg parsing.

This adds a test that fails before and passes after.

I've confirmed that the other mechanisms have tests for int-as-float,
and thus only config files needed a change:

- env var:
https://github.com/pantsbuild/pants/blob/459d39bb39f1f6bef605fd3b4fa129c5d721ea6f/src/rust/engine/options/src/env_tests.rs#L130
- arg:
https://github.com/pantsbuild/pants/blob/459d39bb39f1f6bef605fd3b4fa129c5d721ea6f/src/rust/engine/options/src/args_tests.rs#L107


Fixes #21264

Co-authored-by: Huon Wilson <huon@exoflare.io>