Squash errors from native options parser, treat as discrepancy (Cherry-pick of #21263) (#21272)
This adjusts the code that compares the legacy and native parsers to
catch any exception from the native parser and treat that as a
discrepancy between legacy and native in the same way as the option
values being parsed differently. That is, if the native parser throws an
exception, catch & log it with all the extra info about the legacy ->
native parser switch, rather than letting it bubble up as a normal
exception.

Fixes #21262.

NB. this doesn't fix the underlying int vs. float parsing discrepancy
(see #21264 and #21265), that'll just now be surfaced as logged errors
rather than an exception after this change.

This also makes a few drive-by fixes related to this validation:

- fix the spacing in the docs of
https://www.pantsbuild.org/2.22/reference/global-options#native_options_validation,
so that the list is not softwrapped into oblivion
- log the error/warning as one big message, so that it's (arguably)
easier to understand.
- add some more links
- use the `[GLOBAL]` section name not `[]` when the scope is empty

Before:

```
Exception: Expected [GLOBAL] pantsd_timeout_when_multiple_invocations to be a float but given 1
```

After the basic fix f152565 to convert the hard exception into logged
errors:

```
11:39:00.96 [ERROR] Failed to parse options in scope [] with native parser due to error: Expected [GLOBAL] pantsd_timeout_when_multiple_invocations to be a float but given 1
11:39:00.96 [ERROR] If you can't resolve this discrepancy, please reach out to the Pants development team: https://www.pantsbuild.org/community/getting-help. 
11:39:00.96 [ERROR] The native parser will become the default in 2.23.x, and the legacy parser will be removed in 2.24.x. So it is imperative that we find out about any discrepancies during this transition period.
11:39:00.96 [ERROR] You can use the global native_options_validation option to configure this check.
```

After all the other changes:

```
12:03:02.33 [ERROR] Found differences between the new native options parser and the legacy options parser in scope [GLOBAL]:

- Failed to parse options with native parser due to error:
    Expected [GLOBAL] pantsd_timeout_when_multiple_invocations to be a float but given 1

If you can't resolve this discrepancy, please reach out to the Pants development team: https://www.pantsbuild.org//community/getting-help.

The native parser will become the default in 2.23.x, and the legacy parser will be removed in 2.24.x. So it is imperative that we find out about any discrepancies during this transition period.

You can use the global native_options_validation option (https://www.pantsbuild.org/2.23/reference/global-options#native_options_validation) to configure this check.
```

Co-authored-by: Huon Wilson <huon@exoflare.io>