allow plugins to provide "auxiliary" goals (and refactor BSP into backends) (#20913)
Allow plugins to provide "auxiliary" goals via the newly-introduced
`auxiliary_goals` function in the plugin registration module. This
function returns an iterable of `AuxiliaryGoal` subclasses representing
auxiliary goals which are run outside of the engine.

The first user for this mechanism is the Build Server Protocol support
which previously required the `experimental-bsp` goal to be a
`BuiltinGoal` and thus live in the Pants core rules. With the new
auxiliary goal support, the BSP rules can now live in several new
optional backends and not in the core rules.

See related discussion in
https://github.com/pantsbuild/pants/discussions/20931.

The `BuiltinGoal` class intentionally remains for the "builtin goals" in
the core like the help system. `AuxiliaryGoal` uses a context class
instead of function parameters like `BuiltinGoal` for its `run` method
in order to be more extensible in the future.

---------

Co-authored-by: Jacob Floyd <cognifloyd@gmail.com>