update-build-files: add support for Buildifier formatter (#21208)
`buildifier` was added to the list of supported formatters that can be
used to format the BUILD files with the `update-build-files` goal.

The addition of the Buildifier support was added in
https://github.com/pantsbuild/pants/pull/16573/files; this PR adds a
relevant option to actually let users choose it to format their BUILD
files.

It may be helpful if your organization is migrating from Bazel and wants
to keep the style of the BUILD files
consistent or if for any other reason you may want to adopt the
formatting style that is enforced by `buildifier`.

The `buildifier` can be used on its own, but it can also be used in pair
with a Python formatter, such as `black`
or `ruff`. For instance, you could first run `buildifier` to sort the
target fields alphabetically,
and then run `black` to keep the style consistent with the rest of the
Python code.

E.g. running

```
$ pants update-build-files --fmt --formatter=buildifier src/python/pants/backend/python/util_rules/BUILD
$ pants update-build-files --fmt --formatter=black src/python/pants/backend/python/util_rules/BUILD
```

would convert 

```
resource(name="complete_platform_pex_test", source="complete_platform_pex_test.json")

python_sources(
    overrides={
        "local_dists_pep660.py": dict(dependencies=["./scripts/pep660_backend_wrapper.py"]),
        },
    name="foo",
)
```

to

```
# adding a trailing comma, putting one argument per line
resource(
    name="complete_platform_pex_test",
    source="complete_platform_pex_test.json",
)

# reorder the fields, putting the name first
python_sources(
    name="foo",
    overrides={
        "local_dists_pep660.py": dict(dependencies=["./scripts/pep660_backend_wrapper.py"]),
    },
)
```