Apply InteractiveProcess restartability correctly (#21303)
The switch in #20874 to exposing intrinsics as python functions
meant that running an InteractiveProcess now occurs under a
separate engine Task than the goal that invokes it.

This introduced a bug where the `restartable` property was
applied to the InteractiveProcess intrinsic's Task, and not, as
previously, to the goal's Task.

This change ensures that the restartability applies correctly,
via new rule helpers.

Fixes #21243.