Run all vcs versioning processes locally (#21353)
When using environments we must ensure we run VCS-related
processes in the local environment and not the inherited one,
as a docker environment will not have access to Git state.

In #21188 we made setuptools_scm itself run locally,
but this was not sufficient: this change makes sure we
also run the pex building and git binary discovery
in the local environment.

Fixes #21178