typescript: handle tsconfig.json files that are in non-standard JSON format (#21345)
The `tsconfig.json` file is a configuration file used by the TypeScript
compiler to specify the root files and the compiler options required to
compile a TypeScript project. While it uses a format similar to JSON,
there are some important differences and additional features tailored to
TypeScript projects.

* tsconfig.json allows both single-line (`// comment`) and multi-line
comments (`/* comment */`) to be added
    anywhere in the file.
* Trailing commas in arrays and objects are permitted.

TypeScript uses its own parser to read the file; in standard JSON,
trailing commas or comments are not allowed. This PR makes it possible
to parse these files properly as using plain `json.load` on a JSON file
with comments will result in an error.

---

Not adding any docs or notes since I guess this should be considered to
be default behavior.