A execute_process_or_raise rule alias. (#21375)
`fallible_to_exec_result_or_raise` directly converts a
FallibleProcessResult to a ProcessResult, or raises an
exception if the process failed. Its name makes sense
when you already have a FallibleProcessResult in hand.

But, it is common to want to execute a process and 
automatically raise  an exception on process error.
The new execute_process_or_raise() alias facilitates this idiom:

```
result = await execute_process_or_raise(
    **implicitly(
        Process(...) # Or something that some other rule can convert to a Process.
    )
)
```

Where the execute_process() intrinsic is invoked implicitly to create a
FallibleProcessResult. This is simply a better name for the same rule,
when invoked in this use case.