adhoc/shell: check whether or not outputs were actually captured (#21277)
As reported in https://github.com/pantsbuild/pants/issues/18183, the
`adhoc_tool` and `shell_command` target types do not error when no
outputs are captured nor when a glob specified in`output_files` or
`output_directories` fails to match. This causes confusion for users
because they clearly expected Pants to capture an output and Pants
should not silently accept the situation.

Add a new field `outputs_match_mode` to allow users to configure to what
extent Pants checks the presence of the outputs required by the
`output_files` and `output_directories` fields.