announce (and warn about) end of Python 2.7 support starting in v2.24.x (#21387)
Announce that v2.23.x is the last Pants release series to proactively
test that it works with Python 2.7.

Also, log a warning when Python 2 is detected in the default interpreter
constraints or in an `interpreter_constraints` field. The warning can be
disabled in case the user *really* knows what they want via the new
`--no-python-warn-on-python2-usage` option.