python: a mandatory option is not really "advanced" (interpreter_constraints) (#21397)
Pants currently requires users to set `--python-interpreter-constraints`
and raises an exception if they do not. The option, however, is marked
"advanced" which really should only apply to optional configuration with
defaults most users will not touch, not to mandatory options with no
default.

Remove the "advanced" marking so that this option is more discoverable
by being output for `./pants help python`.