upgrade GitHub Actions image used for most CI jobs and remove Python 2.7 CI testing (#21396)
As part of the [Tonic / Hyper 1.0
upgrade](https://github.com/pantsbuild/pants/pull/21331), we need to
upgrade to a newer version of the Ubuntu image used in CI (from 20.04 to
at least 22.04) due to a transitive dependency requiring a newer C
compiler.

Upgrade the CI image used for most CI builds to `ubuntu-22.04`. This is
currently the same image as `ubuntu-latest` but pins the Ubuntu
distribution used explicitly. (The distribution is pinned to 22.04 for
the ARM runner image, so we do not want the x86 jobs to diverge.)

Python 2.7 was removed from the newer Ubuntu image, however. To avoid
having to re-add Python 2.7 to the CI jobs (since it is well past its
1/1/2020 end of life date), we announced v2.24.x as no longer being
tested with Python 2.7. Consequently, this PR also removes any uses of
Python 2.7 from (CI) testing.