adhoc/shell: allow cache scope to be configured for process executions (#21245)
Expose a subset of the `ProcessCacheScope` capability to users of the
`adhoc_tool` / `shell_command` target types. The new `cache_scope` field
on both target types configures the following cache scope behaviors:

- `from_environment`: Use the default cache scope for the applicable
environment in which the process will execute. (`success` generally
except `session` for `experimental_workspace_environment`). This is the
default.

- `success`: Cache successful executions of the process.

- `success_per_pantsd_restart`: Cache successful executions of the
process for the life of the applicable pantsd process.

- `session`: Only cache the result for a single Pants session. This will
usually be a single invocation of the pants tool.

The intent is to support users who wish to reexecute certain processes
more frequently (probably because of non-idempotent and/or non-pure
semantics of their process).