introduce `SubjectPath` to namespace paths between workspace and system (#21403)
As part of solving https://github.com/pantsbuild/pants/issues/20361,
introduce a new enum `SubjectPath` to the engine representing a
namespaced path (which currently can either be a relative to the
buildroot or a system path).

The goal is to force rule code examining system paths (in to-come PRs)
to make that intention explicitly clear. Rule code should not escape the
buildroot confinement unless it intends to do just that. It also allows
future PRs to track changes on system paths to easily know when such
paths are being requested.

Note: This is a safety issue and not really a security one since rule
code can already escape the buildroot by using `open` and other
filesystem operations directly.