upgrade mypy to v1.10.1 (#21158)
Upgrade MyPy to v1.10.1 for Pants' own source code.

Overview of issues found:
- Many `# type: ignore` comments were removed due to improvements in
MyPy.
- MyPy wonderfully found uses of `dataclasses.replace` where the code
was using the wrong type for a replaced field (especially egregious
since some cases passed in a mutable list where an immutable value was
expected to be stored in the (frozen) dataclass).
- MyPy could not infer that `FrozenDict` implements the
`SupportsKeysAndGetItem` protocol (for whatever reasons) and so I added
ignores of `dict-item` in certain places.
- Upgraded `strawberry-graphql` which is used by repo-internal plugin
`pants-plugins/pants_explorer/server` to work around an ICE in `mypy`. I
also had to add some `# type: ignore` comments to that module. I have
not tested `pants_explorer` since I did not find a README about how to
use it for Pants development.