Replace macOS 10.15 x86-64 self-hosted runner with 12 GH-hosted (#21417)
This removes our use of the self-hosted 10.15 x86-64 runner in favour of
a GitHub-hosted macOS 12 runner, implicitly changing our macOS version
support in the process.

Note: these macOS 12 runners are disappearing in #21333, so I think 2.24
will have to be the last release supporting macOS 12 "properly", and
hopefully we'll have finished that release series before GitHub gets rid
of them.

An alternative that doesn't limit us to GitHub's timelines would be
upgrading OS on the self-hosted runner. I'm disinclined from the
alternative. Landing this PR as-is theoretically allows us to turn off
the self-hosted runner and save money.

Half of #21413