golang: fix exception when coverage enabled and external tests import the base package (Cherry-pick of #21452) (#21454)
As reported in https://github.com/pantsbuild/pants/issues/21386, Pants
raises an exception when coverage is enabled and a package has external
tests (i.e., "xtests") which import the base package. The bug occurred
because the code which injects the base package dependency did not
properly pass through the coverage flag. Consequently, the compile graph
ended up with two different variants of the base package: one with
coverage codegen and the other without coverage codegen. This resulted
in two different copies of the package archive which Pants cannot merge
together into a single `Digest`.

The solution is to properly pass through the coverage flag when
injecting the base package dependency on the external test package to
ensure there are not multiple variants of the base package in the
compile graph.

Fixes #21386.

---------

Co-authored-by: Tom Dyas <tom@shoalsoftware.com>
Co-authored-by: Tom Dyas <tom.dyas@gmail.com>