Make ExternalTool exportable (#21258)
This MR wires `ExternalTool`s into the export machinery.

It exposes them under a separate cli arg `--bin`. Although it uses some
of the same machinery as `--resolve`, there are several differences that
I think support a separate flag (and some separate internal):
1. `ExternalTool`s don't have a resolve. They therefore must not show up
for generating lockfiles. We have to implement this separation
interally, so we should probably surface that.
2. They would be invoked directly (instead of `source
dist/export/.../activate`)
3. We can put them all in a folder "dist/export/bin" so people can get
them all

Closes #21251 

Bikeshedding:
- [x] Do we like `--bin` (yes)
- [x] Is putting all the bins in the same dir what we want (yes, we want
a "bin" folder with all the binaries. But we have to put each binary's
digest in its own folder to prevent collisions of supporting files)
- [-] Do we want to make all TemplatedExternalTools exportable? We could
extend Subsystem.rules and add the UnionRule there (deferred from this
MR)

---------

Co-authored-by: Huon Wilson <wilson.huon@gmail.com>