remove category label requirement from PRs
The category label used to be used as part of release notes
generation, but now that we a spiffy new releases notes process
<https://github.com/pantsbuild/pants/discussions/20888> and they are
no longer used for "major" releases.

They are still part of the github patch releases notes but:
 * Bugfix releases ought to be all bug fixes already
 * It is a point of friction for contributors; every first time
 contributor is likely to get a red X and it is hard for reviewers to
 tell at a glance from the PR list which failures are real.

So on net, the benefit is no longer worth the friction.