"open" support for experimental-explorer on Linux
`open` isn't a traditional Linux command, but
[xdg-open](https://wiki.archlinux.org/title/Default_applications) is a
common part of many distros and exits to "open a file or URL in the
user's preferred application".

NOTE: The intend here is limited to make experimental-explorer
runnable without crashing on Linux, I have not done any testing of how
various GUI applications work or not when invoked from the sandbox.