tests: make cas testutil functions async (#21580)
During the most recent Tonic upgrade, there was a need to use Tokio's
`TcpListener` instead of `std::net::TcpListener` because that is what is
acccepted by Tonic. Unfortunately, using Tokio's `TcpListener` directly
at that time would have required making several CAS testutil functions
be `async` with the corresponding mass refactor changes to call `.await`
on the returned `Future`'s.

I opted to not do those async / await changes at the time to avoid
making the PRs longer to review. Instead, the Tonic refactor created a
`std::net::TcpListener` and wrapped it with the Tokio `TcpListener`.
This is prone to error because the developer needs to remember to make
the `std::net::TcpListener` non-blocking. But it worked to keep the
upgrade PR simpler.

This PR now performs the refactor which was deferred during the Tonic
upgrade.