update Pants' own Python to 3.11
Building on the work in
https://github.com/pantsbuild/scie-pants/pull/351/ this brings the
version of Python used by Pants from 3.9 to 3.11.  Why 3.11 and
not 3.12 or 3.13?  Because that is what we already released on the
scie-pants side and two release forward is still a big benefit.

NOTE: I'd like to hold off on stomping out all deprecation warnings
until one `dev` release since the release process part is what I'm
most nervous about.