use an `EnvironmentTarget` helper instead of direct `isinstance` checks (#21586)
A user reported that Pants was generating a warning for the visibility lint rules because the target was running in an `experimental_workspace_environment`. As it turns out, several parts of the code were checking `EnvironmentTarget.val` directly to see if it was an instance of `LocalEnvironmentTarget`. This is an anti-pattern given that we can introduce new environment types.

This PR fixes the issue by using an existing helper method on `EnvironmentTarget`. The helper method is renamed to `can_access_local_system_paths` from `can_use_system_path_metadata_requests`.